requirements.txt
python-dotenv
python-docx
Jinja2
PyYAML
pyTelegramBotAPI